requirements.txt
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
pandas>=2.0.0
pytest>=7.3.0
flask>=3.0.0
